Finished up tests
Opted to remove the AppDriverTests as it would require high lvl child processing, which has not been taught.

Unable to find a concise answer on how to handle the approach, after not finding a good answer, I asked chatgpt which after a quick through is a very much not in my use case solution.

So I opted to not test this class for now